Initialize repo.
Save registers not implemented.

There is a serious flaw (for QtCreator 3.6.0): Core::EditorManager::hideEditorStatusBar is implemented in such a way that hides the status bar only in current editor. In plugin code this function calls AFTER current editor changed. Thus status bar keep shown in previous editor. The only solution do not show the status bar at all. But it is not clear when the plugin captures key events that can be mistaken for editor hanging. So it requires an alternative approach.